requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
absl-py=1.4.0=pypi_0
appnope=0.1.3=pypi_0
asttokens=2.2.1=pypi_0
astunparse=1.6.3=pypi_0
backcall=0.2.0=pypi_0
blas=2.116=openblas
blas-devel=3.9.0=16_osxarm64_openblas
bzip2=1.0.8=h620ffc9_4
c-ares=1.19.1=hb547adb_0
ca-certificates=2023.5.7=hf0a4a13_0
cachetools=5.3.0=pypi_0
cairo=1.16.0=h29d4eff_2
certifi=2023.5.7=pypi_0
charset-normalizer=3.1.0=pypi_0
comm=0.1.3=pypi_0
debugpy=1.6.7=pypi_0
decorator=5.1.1=pypi_0
eigen=3.4.0=hc021e02_0
executing=1.2.0=pypi_0
ffmpeg=4.2.2=h04105a8_0
flatbuffers=23.5.26=pypi_0
fontconfig=2.13.94=heb65262_0
freetype=2.10.4=h17b34a0_1
gast=0.4.0=pypi_0
gettext=0.21.1=h0186832_0
giflib=5.2.1=h1a8c8d9_3
glib=2.69.1=h514c7bf_2
gmp=6.2.1=h9f76cd9_0
gnutls=3.6.15=h887c41c_0
google-auth=2.19.0=pypi_0
google-auth-oauthlib=1.0.0=pypi_0
google-pasta=0.2.0=pypi_0
graphite2=1.3.14=hc377ac9_1
grpcio=1.54.2=pypi_0
gst-plugins-base=1.14.1=h313beb8_1
gstreamer=1.14.1=h80987f9_1
h5py=3.8.0=pypi_0
harfbuzz=4.3.0=he9eebac_1
hdf5=1.12.1=nompi_had0e5e0_100
icu=68.2=hbdafb3b_0
idna=3.4=pypi_0
ipykernel=6.23.1=pypi_0
ipython=8.13.2=pypi_0
jax=0.4.10=pypi_0
jedi=0.18.2=pypi_0
jpeg=9e=h1a8c8d9_3
jupyter-client=8.2.0=pypi_0
jupyter-core=5.3.0=pypi_0
keras=2.12.0=pypi_0
krb5=1.19.3=hf9b2bbe_0
lame=3.100=h1a8c8d9_1003
lerc=3.0=hc377ac9_0
libblas=3.9.0=16_osxarm64_openblas
libcblas=3.9.0=16_osxarm64_openblas
libclang=16.0.0=pypi_0
libcurl=7.87.0=h0f1d93c_0
libcxx=16.0.4=h4653b0c_0
libdeflate=1.17=h1a8c8d9_0
libedit=3.1.20191231=hc8eb9b7_2
libev=4.33=h642e427_1
libffi=3.4.4=hca03da5_0
libgfortran=5.0.0=12_2_0_hd922786_31
libgfortran5=12.2.0=h0eea778_31
libiconv=1.17=he4db4b2_0
libidn2=2.3.4=h1a8c8d9_0
liblapack=3.9.0=16_osxarm64_openblas
liblapacke=3.9.0=16_osxarm64_openblas
libllvm12=12.0.1=h93073aa_2
libnghttp2=1.46.0=h95c9599_0
libopenblas=0.3.21=openmp_hc731615_3
libopus=1.3.1=h27ca646_1
libpng=1.6.39=h80987f9_0
libpq=12.9=h65cfe13_3
libssh2=1.10.0=hb80f160_2
libtasn1=4.19.0=h1a8c8d9_0
libtiff=4.5.0=h313beb8_2
libunistring=0.9.10=h3422bc3_0
libvpx=1.10.0=hc377ac9_0
libwebp=1.2.4=ha3663a8_1
libwebp-base=1.2.4=h80987f9_1
libxml2=2.10.3=h372ba2a_0
libxslt=1.1.37=h1bd8bc4_0
llvm-openmp=16.0.4=h1c12783_0
lz4-c=1.9.4=hb7217d7_0
markdown=3.4.3=pypi_0
markupsafe=2.1.2=pypi_0
matplotlib-inline=0.1.6=pypi_0
ml-dtypes=0.1.0=pypi_0
ncurses=6.4=h313beb8_0
nest-asyncio=1.5.6=pypi_0
nettle=3.7.3=h84b5d62_1
nspr=4.35=hb7217d7_0
nss=3.74=h142855e_0
numpy=1.23.5=pypi_0
numpy-base=1.24.3=py311h1d85a46_0
oauthlib=3.2.2=pypi_0
openblas=0.3.21=openmp_hf78f355_3
opencv=4.6.0=py311h6956b77_3
openh264=1.8.0=h98b2900_0
openssl=1.1.1t=h03a7124_0
opt-einsum=3.3.0=pypi_0
packaging=23.1=pypi_0
pandas=2.1.4=pypi_0
parso=0.8.3=pypi_0
pathspec=0.12.1=pypi_0
pcre=8.45=hbdafb3b_0
pexpect=4.8.0=pypi_0
pickleshare=0.7.5=pypi_0
pip=23.0.1=py311hca03da5_0
pixman=0.40.0=h27ca646_0
platformdirs=3.5.1=pypi_0
prompt-toolkit=3.0.38=pypi_0
protobuf=4.23.2=pypi_0
psutil=5.9.5=pypi_0
psycopg2-binary=2.9.9=pypi_0
ptyprocess=0.7.0=pypi_0
pure-eval=0.2.2=pypi_0
pyasn1=0.5.0=pypi_0
pyasn1-modules=0.3.0=pypi_0
pygments=2.15.1=pypi_0
python=3.11.3=hc0d8a6c_0
python-dateutil=2.8.2=pypi_0
pytz=2023.3.post1=pypi_0
pyyaml=6.0.1=pypi_0
pyzmq=25.0.2=pypi_0
qt-main=5.15.2=ha2d02b5_7
qt-webengine=5.15.9=h2903aaf_4
qtwebkit=5.212=h19f419d_5
readline=8.2=h1a28f6b_0
requests=2.31.0=pypi_0
requests-oauthlib=1.3.1=pypi_0
rsa=4.9=pypi_0
scipy=1.10.1=pypi_0
setuptools=66.0.0=py311hca03da5_0
six=1.16.0=pypi_0
sqlalchemy=2.0.25=pypi_0
sqlite=3.41.2=h80987f9_0
stack-data=0.6.2=pypi_0
tensorboard=2.12.3=pypi_0
tensorboard-data-server=0.7.0=pypi_0
tensorflow-estimator=2.12.0=pypi_0
tensorflow-macos=2.12.0=pypi_0
tensorflow-metal=0.8.0=pypi_0
termcolor=2.3.0=pypi_0
tk=8.6.12=hb8d0fd4_0
tornado=6.3.2=pypi_0
traitlets=5.9.0=pypi_0
typing-extensions=4.6.2=pypi_0
tzdata=2023.4=pypi_0
urllib3=1.26.16=pypi_0
wcwidth=0.2.6=pypi_0
werkzeug=2.3.4=pypi_0
wheel=0.38.4=py311hca03da5_0
wrapt=1.14.1=pypi_0
x264=1!152.20180806=h1a28f6b_0
xz=5.4.2=h80987f9_0
yamllint=1.33.0=pypi_0
zlib=1.2.13=h5a0b063_0
zstd=1.5.5=hd90d995_0
